adapted to moved PCMSolverWorkflowRunConfiguration
