+++ bundles/org.palladiosimulator.reliability.solver/src/org/palladiosimulator/reliability/solver/runconfig/ReliabilityWorkflowJobBuilder.java
@@ -3,7 +3,7 @@ package org.palladiosimulator.reliability.solver.runconfig;
 import org.eclipse.debug.core.ILaunch;
 import org.palladiosimulator.analyzer.workflow.core.configurations.AbstractPCMWorkflowRunConfiguration;
 import org.palladiosimulator.analyzer.workflow.jobs.PCMWorkflowJobBuilder;
-import org.palladiosimulator.solver.runconfig.PCMSolverWorkflowRunConfiguration;
+import org.palladiosimulator.solver.core.runconfig.PCMSolverWorkflowRunConfiguration;
 
 import de.uka.ipd.sdq.workflow.jobs.IJob;
